requirements.txt
Flask==1.0.2
Keras==2.2.2
numpy==1.14.5
opencv-python==3.4.2.17
Pillow==5.2.0
requests==2.19.1
tensorflow==1.10.0
imutils==0.5.0
scikit-learn
pytesseract